Raytrace some spheres
For every point on the canvas calculate if it hits one of the spheres or not to determine the color of the pixel.

Loads of boiler plating went in with the vectors and colors and the math for those.

Did a DrawRect in another program so copied it in here.

Did some clean up of code that isn't getting used.

Clear the screen to black and draw my bitmap rougly centered.

Start the program in fullscreen.