Basic lighting added to raytracer
Raytracer now does a terrible job with lighting, but it can be ambient, from a specific place, or from the sun. It can be diffuse or shiny.

My truncate function was actually a clamp.

My color functions now truncate, but can use floats.

Now using a really bad power function for floats.