Trying to figure out what's wrong.
So I went back to the previous version of the code, but fixed specular lighting. It has the exact same error to reflections, only the colors are inverted. Super crazy. I also noticed that my reflections aren't quite the same as the book I'm following and my spheres are brighter than they should be. So something is off with my math. I don't know if I want to fix it right now.